+++ src/admin/loadContactsTable.js
@@ -2,24 +2,30 @@ import { getDataContacts } from "../componentes/contacto.js";
 
 window.addEventListener('load', async () => {
     await crearTablaContacts();
-    document.querySelectorAll(".button-Editar").forEach(button => {
-        button.addEventListener('click', (event) => {
-            const contactId = event.target.id;
-            location.href = `/pages/admin/editContact.html?contactId=${contactId}`;
+    const sessionToken = sessionStorage.getItem('accessToken');
+    if (sessionToken !== undefined && sessionToken !== null && sessionToken !== 'null') {
+        document.querySelectorAll(".button-Editar").forEach(button => {
+            button.addEventListener('click', (event) => {
+                const contactId = event.target.id;
+                location.href = `/pages/admin/editContact.html?contactId=${contactId}`;
+            });
         });
-    });
 
-    document.querySelectorAll(".button-Eliminar").forEach(button => {
-        button.addEventListener('click', async (event) => {
-            const contactId = event.target.id;
-            const userResponse = confirm('Esta seguro en eliminar el Contacto');
-            if (userResponse) {
-                await deleteContact(contactId);
-                alert('El contacto ha sido eliminado correctamente.');
-                location.href = '/pages/admin/contactsAdmin.html';
-            }
+        document.querySelectorAll(".button-Eliminar").forEach(button => {
+            button.addEventListener('click', async (event) => {
+                const contactId = event.target.id;
+                const userResponse = confirm('Esta seguro en eliminar el Contacto');
+                if (userResponse) {
+                    await deleteContact(contactId);
+                    alert('El contacto ha sido eliminado correctamente.');
+                    location.href = '/pages/admin/contactsAdmin.html';
+                }
+            });
         });
-    });
+    } else {
+        alert('Por favor Iniciar Sesión');
+        location.href = '../login.html';
+    }
 });
 
 const crearTablaContacts = async () => {
@@ -122,15 +128,13 @@ const buttonComponent = (buttonLabel, id) => {
 
 const deleteContact = async (id) => {
     try {
-        /*const sessionToken = sessionStorage.getItem('accessToken');*/
-        let response = await fetch(`http://localhost:3000/api/contacts/${id}`, 
+        let response = await fetch(`http://localhost:3000/api/contacts/${id}`,
             {
                 method: 'DELETE',
                 headers: {
                     'Content-Type': 'application/json',
-                    /*'Authorization': `${sessionToken}`*/
                 },
-        });
+            });
         return response;
     } catch (error) {
         console.error('Hubo un error');
